+++ src/eduid/scimapi/context.py
@@ -3,12 +3,12 @@ import logging.config
 from typing import Optional, Union
 from uuid import UUID
 
+from eduid.common.fastapi.log import init_logging
 from eduid.common.models.scim_base import SCIMResourceType
 from eduid.common.utils import urlappend
 from eduid.queue.db.message import MessageDB
 from eduid.scimapi.config import DataOwnerName, ScimApiConfig
 from eduid.scimapi.context_request import ContextRequest
-from eduid.scimapi.log import init_logging
 from eduid.scimapi.notifications import NotificationRelay
 from eduid.scimapi.utils import load_jwks, make_etag
 from eduid.userdb.scimapi import ScimApiEventDB, ScimApiGroup, ScimApiGroupDB
